Use -R instead of -r in kubedep.sh
Today I learned: On MacOS, -r and -R are different flags, and -r is a legacy
flag that doesn't work with symbolic links.  It even conflicts with the -L
flag:

```
cp: the -H, -L, and -P options may not be specified with the -r option.
```

On linux they're the same though.